+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * **Breaking Change:** cli arguments `--fatal-unused` and `--fatal-warnings` activate by default.
 * **Breaking Change:** rename rules prefixed with `no-` to use `avoid-` prefix.
+* feat: add [prefer-immediate-return](https://dartcodemetrics.dev/docs/rules/common/prefer-immediate-return) rule
+* feat: add Checkstyle format reporter.
 * chore: restrict `analyzer` version to `>=3.0.0 <3.4.0`.
 * chore: restrict `analyzer_plugin` version to `>=0.9.0 <0.10.0`.